add all
add Dockerfile

modify some error

add airflow

modif word
