+++ integration_tests/iceberg-sink/airflow_dags/rewrite_iceberg_small_files.py
@@ -13,7 +13,7 @@ with DAG(
     start_date=pendulum.datetime(2021, 1, 1, tz="UTC"),
     catchup=False,
     dagrun_timeout=datetime.timedelta(minutes=60),
-    schedule_interval= datetime.timedelta(minutes=5),
+    schedule_interval= datetime.timedelta(hours=3),
     tags=["iceberg"],
 ) as dag:
